fix: Delete unuse code
